Fix saving of host comments in Zenmap. Fixes #133
Sometimes (all the time?) hosts are "cloned" or deep-copied instead of being
copied by reference, so testing whether a host is "in" a scan is insufficient.
Now we check by address, the same way that Zenmap uses to determine 2 hosts are
the "same" for merging their output.

Also, avoid considering a scan as having unsaved changes if the comment field is
merely focused but unchanged.


git-svn-id: https://svn.nmap.org/nmap@34537 e0a8ed71-7df4-0310-8962-fdc924857419